Adding new fields to voyage model.
